fix: P4ADEV-3829 CVE fix + fix regression introduced updating SpringDataJpa module

+++ src/main/java/it/gov/pagopa/pu/classification/controller/TreasuryEntityExtendedController.java
@@ -24,7 +24,7 @@ public class TreasuryEntityExtendedController implements TreasuryEntityExtendedC
   }
 
   @Override
-  public ResponseEntity<Long> deleteByOrganizationIdAndBillCodeAndBillYearAndOrgBtCodeAndOrgIstatCode(Long organizationId, String billCode, String billYear, String orgBtCode, String orgIstatCode){
+  public ResponseEntity<Integer> deleteByOrganizationIdAndBillCodeAndBillYearAndOrgBtCodeAndOrgIstatCode(Long organizationId, String billCode, String billYear, String orgBtCode, String orgIstatCode){
     return ResponseEntity.ok(repository.deleteByOrganizationIdAndBillCodeAndBillYearAndOrgBtCodeAndOrgIstatCode(organizationId, billCode, billYear, orgBtCode, orgIstatCode));
   }
 }
